Moced create ai match to playground.py from goMCTS.py
